# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  profile: 'profile',
  document: 'document',
  eligibility: 'eligibility',
  application: 'application',
  verification: 'verification',
} as const;

export interface Activity {
  id: string;
  title: string;
  description: string;
  timestamp: string;
  type: ActivityType;
}

export interface DashboardSummary {
  citizenName: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  readinessScore: number;
  eligibleServices: number;
  uploadedDocuments: number;
  activeApplications: number;
  pendingActions: number;
  timeSavedMinutes: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  profileCompletion?: number;
}

export interface CitizenProfile {
  id: string;
  fullName: string;
  email: string;
  mobile: string;
  age: number;
  gender: string;
  aadhaarMasked: string;
  education: string;
  occupation: string;
  annualIncome: number;
  socialCategory: string;
  state: string;
  district: string;
  village: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  completion: number;
}

export interface ProfileUpdate {
  fullName?: string;
  mobile?: string;
  age?: number;
  gender?: string;
  education?: string;
  occupation?: string;
  annualIncome?: number;
  socialCategory?: string;
  state?: string;
  district?: string;
  village?: string;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  verified: 'verified',
  pending: 'pending',
  missing: 'missing',
} as const;

export interface Document {
  id: string;
  name: string;
  category: string;
  status: DocumentStatus;
  verified: boolean;
  updatedAt: string;
  size: string;
  requiredFor?: string[];
}

export interface DocumentInput {
  name: string;
  category: string;
  size?: string;
}

export type GovernmentServiceStatus = typeof GovernmentServiceStatus[keyof typeof GovernmentServiceStatus];


export const GovernmentServiceStatus = {
  eligible: 'eligible',
  'needs-action': 'needs-action',
  ineligible: 'ineligible',
} as const;

export interface GovernmentService {
  id: string;
  name: string;
  category: string;
  description: string;
  status: GovernmentServiceStatus;
  processingTime: string;
  requiredDocuments: string[];
  reason?: string;
  accent?: string;
}

export type EligibilityResultStatus = typeof EligibilityResultStatus[keyof typeof EligibilityResultStatus];


export const EligibilityResultStatus = {
  eligible: 'eligible',
  'needs-action': 'needs-action',
  ineligible: 'ineligible',
} as const;

export interface EligibilityResult {
  serviceId: string;
  serviceName: string;
  status: EligibilityResultStatus;
  reason: string;
  missingDocuments: string[];
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  'in-review': 'in-review',
  'action-needed': 'action-needed',
  approved: 'approved',
  submitted: 'submitted',
} as const;

export type ApplicationStepStatus = typeof ApplicationStepStatus[keyof typeof ApplicationStepStatus];


export const ApplicationStepStatus = {
  complete: 'complete',
  current: 'current',
  upcoming: 'upcoming',
} as const;

export interface ApplicationStep {
  label: string;
  status: ApplicationStepStatus;
  /** @nullable */
  timestamp?: string | null;
}

export interface Application {
  id: string;
  serviceName: string;
  reference: string;
  status: ApplicationStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  updatedAt: string;
  steps: ApplicationStep[];
}

export interface ApplicationInput {
  serviceId: string;
}

export interface DemoState {
  active: boolean;
  step: number;
  message: string;
}

export type AdminQueueItemPriority = typeof AdminQueueItemPriority[keyof typeof AdminQueueItemPriority];


export const AdminQueueItemPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface AdminQueueItem {
  id: string;
  applicant: string;
  serviceName: string;
  priority: AdminQueueItemPriority;
  submittedAt: string;
}

export interface AdminOverview {
  pendingReview: number;
  verificationRequests: number;
  approvedToday: number;
  averageProcessingTime: string;
  queue: AdminQueueItem[];
}

export type GetServicesParams = {
category?: string;
search?: string;
};