# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get citizen profile
      responses:
        "200":
          description: Citizen profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitizenProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update citizen profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated citizen profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitizenProfile"
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: List document vault items
      responses:
        "200":
          description: Document vault
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Add document metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: List government services
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Government service marketplace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovernmentService"
...
  /applications:
    get:
      operationId: getApplications
      tags: [applications]
      summary: List citizen applications
      responses:
        "200":
          description: Application tracking items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Start a service application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
components:
  schemas:
...
    CitizenProfile:
      type: object
      required: [id, fullName, email, mobile, age, gender, aadhaarMasked, education, occupation, annualIncome, socialCategory, state, district, village, completion]
      properties:
        id: { type: string }
        fullName: { type: string }
        email: { type: string, format: email }
        mobile: { type: string }
        age: { type: integer }
        gender: { type: string }
        aadhaarMasked: { type: string }
        education: { type: string }
        occupation: { type: string }
        annualIncome: { type: integer }
        socialCategory: { type: string }
        state: { type: string }
        district: { type: string }
        village: { type: string }
        completion: { type: integer, minimum: 0, maximum: 100 }
    ProfileUpdate:
      type: object
      properties:
        fullName: { type: string }
        mobile: { type: string }
        age: { type: integer }
        gender: { type: string }
        education: { type: string }
        occupation: { type: string }
        annualIncome: { type: integer }
        socialCategory: { type: string }
        state: { type: string }
        district: { type: string }
        village: { type: string }
    Document:
      type: object
      required: [id, name, category, status, verified, updatedAt, size]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        status:
          type: string
          enum: [verified, pending, missing]
        verified: { type: boolean }
        updatedAt: { type: string }
        size: { type: string }
        requiredFor:
          type: array
          items: { type: string }
    DocumentInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string }
        category: { type: string }
        size: { type: string }
    GovernmentService:
      type: object
      required: [id, name, category, description, status, processingTime, requiredDocuments]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        status:
          type: string
          enum: [eligible, needs-action, ineligible]
        processingTime: { type: string }
        requiredDocuments:
          type: array
          items: { type: string }
        reason: { type: string }
        accent: { type: string }
...
    Application:
      type: object
      required: [id, serviceName, reference, status, progress, updatedAt, steps]
      properties:
        id: { type: string }
        serviceName: { type: string }
        reference: { type: string }
        status:
          type: string
          enum: [in-review, action-needed, approved, submitted]
        progress: { type: integer, minimum: 0, maximum: 100 }
        updatedAt: { type: string }
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationStep"
    ApplicationStep:
      type: object
      required: [label, status]
      properties:
        label: { type: string }
        status:
          type: string
          enum: [complete, current, upcoming]
        timestamp: { type: ["string", "null"] }
    ApplicationInput:
      type: object
      required: [serviceId]
      properties:
        serviceId: { type: string }
...