# lib/api-zod/src/generated/types/applicationStep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStepStatus } from './applicationStepStatus';

export interface ApplicationStep {
  label: string;
  status: ApplicationStepStatus;
  /** @nullable */
  timestamp?: string | null;
}