# lib/api-zod/src/generated/types/citizenProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CitizenProfile {
  id: string;
  fullName: string;
  email: string;
  mobile: string;
  age: number;
  gender: string;
  aadhaarMasked: string;
  education: string;
  occupation: string;
  annualIncome: number;
  socialCategory: string;
  state: string;
  district: string;
  village: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  completion: number;
}