# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: string;
  name: string;
  category: string;
  status: DocumentStatus;
  verified: boolean;
  updatedAt: string;
  size: string;
  requiredFor?: string[];
}