# lib/api-zod/src/generated/types/documentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentInput {
  name: string;
  category: string;
  size?: string;
}