# lib/api-zod/src/generated/types/governmentService.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GovernmentServiceStatus } from './governmentServiceStatus';

export interface GovernmentService {
  id: string;
  name: string;
  category: string;
  description: string;
  status: GovernmentServiceStatus;
  processingTime: string;
  requiredDocuments: string[];
  reason?: string;
  accent?: string;
}