# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  fullName?: string;
  mobile?: string;
  age?: number;
  gender?: string;
  education?: string;
  occupation?: string;
  annualIncome?: number;
  socialCategory?: string;
  state?: string;
  district?: string;
  village?: string;
}